chore: Create a generate
